+++ custom_components/energy_consistency/const.py
@@ -6,7 +6,7 @@ from datetime import timedelta
 
 DOMAIN = "energy_consistency"
 PLATFORMS = ["sensor"]
-FRONTEND_VERSION = "0.7.2"
+FRONTEND_VERSION = "0.7.3"
 
 CONF_OFFICIAL_ENERGY_ENTITY = "official_energy_entity"
 CONF_OFFICIAL_DATE_ENTITY = "official_date_entity"